Pin fog-core
The fog-core gem added a dependency on the xmlrpc gem which depends on
ruby 2.3 which is not available on centos7. It seems like they tried to
avoid this issue but weren't successful at it[1], so we'll blacklist the
current version and assume it will be fixed in the next version.

[1] https://github.com/fog/fog-core/issues/206

Change-Id: I968aa64e958cd3e9567fd47981d50b302960783d