Move spec_helper_acceptance into helper gem
Every puppet module contains the same file that prepares beaker tests.
Currently, when we need to make changes to it, we need to change it in
every puppet module. This is inefficient and causes a huge review
backlog for identical changes. This patch copies the
spec_helper_acceptance.rb file, with minor modifications to keep the
path logic correct, into the common helper gem. The modules already
depend on this gem for common gem dependencies. This way, modules can
simply `require
puppet-openstack_infra_spec_helper/spec_helper_acceptance` in their
tests and not bother keeping the spec_helper_acceptance module locally.

Change-Id: I18ed6cd2f8ee472a66cf580e3e183d14c00e6877