Refactor spec helper
The spec helper file has been cargocult-copied from puppetlabs and
puppet-openstack modules, adjusted slightly for Infra's particular
needs. It is hard to read which makes adjusting it hard to do and hard
to review. This refactor moves code into named functions and condenses
the main setup code so that there is only one loop over the host list.

Change-Id: I572a5dc0d9987e8c90a27f300c8b4979e20c43a2