Cap rake below 12.3.0
New rake version 12.3.0 needs ruby 2.0 or greater which is not available
on our trusty test nodes. Cap to under rake 12.3.0 so that we use older
rake until we can stop using trusty and have more modern ruby.

The test runner script updates the installation path for this helper lib
to point back at itself when running local tests. Unfortunately it was
doing so with a buggy sed that no longer matched the contents of the
Gemfile that needed updating.

Fix this by updating the sed command to properly point to this lib for
local testing. Note this is fixed here because we need both fixes in
together in order to get gate jobs to pass.

Change-Id: Iddbbab9ea5996df4922bf7927deb8f0354378ab7