Fix installation path in test runner script
The test runner script updates the installation path for this helper lib
to point back at itself when running local tests. Unfortunately it was
doing so with a buggy sed that no longer matched the contents of the
Gemfile that needed updating.

Fix this by updating the sed command to properly point to this lib for
local testing.

Change-Id: I2d42d29e58df36ad0efb83901726234bfc03dab7