Check if we are running under Zuul v3
Update the Zuul conditional check to handle Zuul v3. Zuul v3 no longer
sets a URL or BRANCH env variable so we shouldn't rely on those for
checking if we are running under Zuul. Instead we can check if the
ZUUL_UUID is set which will be present for every job.

Depends-On: I83e7088c36b1b57279bb32e44c9f74730da582cf
Change-Id: Id4f6ae8cc4128083f0db72fb6e2dc81b9968e884