Don't be overzealous with passing the environment
Without this change, the gem helper passes the entirely of the user's
environment through beaker's SSH connection. This causes problems with
puppet 4 because all of the ruby and bundler arguments confuse the ruby
interpreter and we see errors like this:

  /var/lib/gems/2.3.0/gems/bundler-1.16.2/lib/bundler/spec_set.rb:91:in `block in materialize'  :   Could not find rake-12.2.1 in any of the sources   (  Bundler::GemNotFound  )

The only important variable is PUPPET_VERSION so just pass that and
nothing else.

Change-Id: I678716890d0c749e755ec0526621d47fee44a4bb